Fixed the origin of the duplicate elements in the collection
After a long time researching what would be causiing the duplication of elements in a collection, I found out that it was because of both the way the join promise was being called (join in line 94 needs a callback with the two parameters for each result instead of a .then with the two results) but that was anyway passing the unit tests silently, if there were fixtures with no "id" forced. As I have forced IDs, when inserted, it inserted two elements in the collection. Added unit tests to check that only one document is inserted. Finally, changed the join promise with a cascaded .then, as it saves having to do a merge, because the seed object is actually passed by reference, so there is no need to merge it.